BUG: track all clauses with at least one literal
Clauses with one literal are assertions that warrant immediate
assignment of a value. Since we knew their values, we weren't tracking
these clause at all. This leads to problems in the Policy, which assumes
that all package literals will appear in at least one clause.

Clauses with no literals are still ignored.