FIX: Avoid infinite loop in policy recommendations
Using the locally installed version of a candidate in the policy works
a lot of the time but can lead to errors.
The policy came up with a candidate to propose. We hijack that and
suggest a different (equivalent) candidate. The next time around, the
policy may try to suggest the original candidate again and we get stuck
in an infinite loop. This would have been possible on before this
change, too, but would not come up because generally prefer_installed
was set to True.

Adding postprocessing in `Transaction` prevents removing an installed
package and then re-installing the same version from another repository.